smart_video_summarizer_backend/requirements.txt
fastapi
uvicorn
python-multipart
scipy
nltk
scikit-learn
numpy
sentencepiece
protobuf
google-generativeai
yt-dlp
youtube-transcript-api
ollama
opencv-python
Pillow
python-dotenv
docx2txt
pymupdf
sqlalchemy
requests
